# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manager API for ShopManagement",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "AllUsersWithOptionalFilters"
                ],
                "summary": "Get all users with optional filters",
                "description": "Récupération des utilisateurs avec possibilité de filtrer par role et/ou par active",
                "operationId": "GetAllUsersWithOptionalFilters",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Filterer les utilisateurs par role (admin ou boutiquier)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "admin",
                                "boutiquier"
                            ]
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Filterer les utilisateurs par status (oui ou non)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "oui",
                                "non"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération reussie",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "StoreUser"
                ],
                "summary": "Store User",
                "description": "Creation d'un compte utilisateur avec le role ADMIN ou BOUTIQUIER",
                "operationId": "StoreUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "photo",
                                    "login",
                                    "password",
                                    "password_confirmation",
                                    "role"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "example": "https://images.pexels.com/photos/633432/pexels-photo-633432.jpeg?auto=compress&cs=tinysrgb&w=600"
                                    },
                                    "login": {
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Passer@123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Passer@123"
                                    },
                                    "role": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "1"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "photo",
                                    "login",
                                    "password",
                                    "password_confirmation",
                                    "role"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "example": "https://images.pexels.com/photos/633432/pexels-photo-633432.jpeg?auto=compress&cs=tinysrgb&w=600"
                                    },
                                    "login": {
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Passer@123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Passer@123"
                                    },
                                    "role": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "1"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}